// src/api/ambulance-ufe/models/CommunicationProtocol.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Remote Examination Api
 * Remote Examination Support management for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: davidtruhlarac@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CommunicationProtocol
 */
export interface CommunicationProtocol {
    /**
     * Unique id of the protocol
     * @type {string}
     * @memberof CommunicationProtocol
     */
    id: string;
    /**
     * Reference to the consultation entry
     * @type {string}
     * @memberof CommunicationProtocol
     */
    entryId: string;
    /**
     * Content of the communication protocol
     * @type {string}
     * @memberof CommunicationProtocol
     */
    content: string;
    /**
     * Current status of the protocol
     * @type {string}
     * @memberof CommunicationProtocol
     */
    status: CommunicationProtocolStatusEnum;
    /**
     * Timestamp of last update
     * @type {Date}
     * @memberof CommunicationProtocol
     */
    updatedAt?: Date;
}


/**
 * @export
 */
export const CommunicationProtocolStatusEnum = {
    Open: 'open',
    InProgress: 'in-progress',
    Closed: 'closed'
} as const;
export type CommunicationProtocolStatusEnum = typeof CommunicationProtocolStatusEnum[keyof typeof CommunicationProtocolStatusEnum];


/**
 * Check if a given object implements the CommunicationProtocol interface.
 */
export function instanceOfCommunicationProtocol(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "entryId" in value;
    isInstance = isInstance && "content" in value;
    isInstance = isInstance && "status" in value;

    return isInstance;
}

export function CommunicationProtocolFromJSON(json: any): CommunicationProtocol {
    return CommunicationProtocolFromJSONTyped(json, false);
}

export function CommunicationProtocolFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommunicationProtocol {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'entryId': json['entryId'],
        'content': json['content'],
        'status': json['status'],
        'updatedAt': !exists(json, 'updatedAt') ? undefined : (new Date(json['updatedAt'])),
    };
}

export function CommunicationProtocolToJSON(value?: CommunicationProtocol | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'entryId': value.entryId,
        'content': value.content,
        'status': value.status,
        'updatedAt': value.updatedAt === undefined ? undefined : (value.updatedAt.toISOString()),
    };
}